Feature: add functionality to toggle rendered icon sizes in demo app (#199)
* enhancement: connect toggle to displayed icon list
