--- ember-flight-icons/tests/dummy/app/components/ds-toggle-button.js
@@ -1,18 +1,0 @@
-import Component from '@glimmer/component';
-import { action } from '@ember/object';
-import { tracked } from '@glimmer/tracking';
-
-export default class DsToggleButtonComponent extends Component {
-  @tracked visibleIconSize = '16';
-
-  @action
-  updateVisibleIconSize() {
-    if (this.visibleIconSize === '16') {
-      this.visibleIconSize = '24';
-    } else {
-      this.visibleIconSize = '16';
-    }
-    // TODO take this out before publishing
-    console.log(`visibleIconSize is ${this.visibleIconSize}`);
-  }
-}
