+++ ember-flight-icons/tests/dummy/app/routes/index.js
@@ -1,26 +1,3 @@
 import Route from '@ember/routing/route';
-import fetch from 'fetch';
-import { getOwner } from '@ember/application';
 
-export default class IndexRoute extends Route {
-  get contextRootURL() {
-    const config = getOwner(this).resolveRegistration('config:environment');
-    return config.rootURL || '/';
-  }
-
-  async model() {
-    const response = await fetch(
-      `${this.contextRootURL}@hashicorp/ember-flight-icons/icons/catalog.json`
-    );
-    const json = await response.json();
-
-    return json.assets.map(({ iconName, fileName, size, description }) => {
-      return {
-        iconName: `${iconName}`,
-        name: `${fileName}`,
-        size: `${size}`,
-        searchable: `${iconName}, ${description}`,
-      };
-    });
-  }
-}
+export default class IndexRoute extends Route {}
