+++ ember-flight-icons/tests/dummy/app/styles/app.css
@@ -37,6 +37,7 @@ main.ds-main {
 }
 
 ul.ds-ul-grid {
+  box-sizing: border-box;
   display: grid;
   font-family: var(--font-body);
   gap: 16px;
@@ -49,39 +50,43 @@ ul.ds-ul-grid {
 }
 ul.ds-ul-grid li.ds-li {
   align-items: center;
-  display: flex;
-  flex-direction: column;
-  font-size: 0.75em;
-  gap: 1ch;
-  justify-content: center;
-  padding: 0.5em 0.15em 0.15em;
   background: var(--gray-7);
   border-radius: 3px;
   box-shadow: 0 0 0 1px var(--gray-5);
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  gap: 1ch;
+  justify-content: center;
+  padding: 0.5em 0.15em 0.15em;
 }
 ul.ds-ul-grid li.ds-li:hover {
   background-color: var(--info-l2);
   box-shadow: 0 0 0 1px var(--info-l1);
 }
 ul.ds-ul-grid li.ds-li .ds-icon-frame {
-  padding: 6px;
-  margin-top: auto;
-  display: flex;
   background: white;
   border-radius: 2px;
   box-shadow: 0 0 0 1px var(--gray-5);
+  box-sizing: border-box;
+  display: flex;
+  margin-top: auto;
+  padding: 6px;
 }
 ul.ds-ul-grid li.ds-li:hover .ds-icon-frame {
-  color: var(--brand-link);
   box-shadow: 0 0 0 1px var(--info-l1);
+  box-sizing: border-box;
+  color: var(--brand-link);
 }
 ul.ds-ul-grid li.ds-li p {
-  text-align: center;
+  box-sizing: border-box;
+  font-size: 0.75em;
   margin-top: auto;
-  width: 100%;
-  text-overflow: ellipsis;
   overflow: hidden;
+  text-align: center;
+  text-overflow: ellipsis;
   white-space: nowrap;
+  width: 100%;
 }
 
 .d-none {
